+++ RELEASE-NOTES.md
@@ -1,8 +1,23 @@
 # DOCMS Release Notes
 
-## yyyy-MM-dd - v0.1.0
+## 2023-04-13
 
-First release:
+CLI `v0.1.1`:
+- Command `new` - helper to create assets with default metadata.
+  - `new site`: Create new site content metadata
+  - `new topic`: Create new topic metadata
+  - `new document`: Create new document metadata
 
-- Feature 1
-- Feature 2
+Runtime `v0.1.1`:
+- Bug fixes.
+
+## 2023-04-12 - v0.1.0
+
+CLI:
+- Pre-process website content with command `docli build`.
+- Build fulltext index while pre-processing website content.
+
+Runtime:
+- Render website content, supporting 3 themes: `bootstrap` (default), `coderdocs` and `prettydocs`.
+- Fulltext search.
+- Multi-languages support.